#25 - chore: 종류 컴포넌트에 Suspense 및 ErrorBoundary 적용
